Update patch to modify download location for lyrics
I've changed the lyrics download path from $XDG_DATA_HOME/lyrics to
$XDG_DATA_HOME/$APP_ID/lyrics instead, because the former path looks way
too common for comfort.

I know that shouldn't matter for Flatpak, but since I'll be trying to
upstream an improved version of this patch later, the old path could
potentially cause issues for non-Flatpak installations of this app.

Lastly, users unlucky enough to run the app with the old path will have
to manually reset the app settings for this new path to be effective
(Hamburger menu > Restore default settings).